Remodeled into single page app
