+++ src/App.js
@@ -7,7 +7,8 @@ import Qs from 'qs';
 import Header from './Components/Header';
 import Main from './Components/Main';
 import Footer from './Components/Footer';
-import SearchResults from './Components/SearchResults';
+
+// import SearchResults from './Components/SearchResults';
 const convert = require('xml-js');
 
 
@@ -23,6 +24,8 @@ class App extends Component {
 
       searchInput: "",
       // User search input for SearchBar
+
+      // searchedBooks: "",
 
       userLibrary: [],
       // User selected from API books
@@ -110,88 +113,85 @@ class App extends Component {
   }
 
 
-  handleFormSubmit = (e) => {
-    e.preventDefault();
+//   handleFormSubmit = (e) => {
+//     e.preventDefault();
 
-    axios ({
-        url: 'https://proxy.hackeryou.com',
+//     axios ({
+//         url: 'https://proxy.hackeryou.com',
 
-        responseType: '',
+//         responseType: '',
 
-// paramsSerializer allows us to pass query params into axios call
-        paramsSerializer: function (params) {
-            return Qs.stringify(params, {arrayFormat: 'brackets'})
-        },
+// // paramsSerializer allows us to pass query params into axios call
+//         paramsSerializer: function (params) {
+//             return Qs.stringify(params, {arrayFormat: 'brackets'})
+//         },
 
-        params: {
-            reqUrl: "https://www.goodreads.com/search/index.xml",
+//         params: {
+//             reqUrl: "https://www.goodreads.com/search/index.xml",
         
-            params : {
-                key: '14csXzY0xdicnCrXfQSO1w',
-                method: "search_index",
-                q: this.state.searchInput,
-            },
+//             params : {
+//                 key: '14csXzY0xdicnCrXfQSO1w',
+//                 method: "search_index",
+//                 q: this.state.searchInput,
+//             },
         
-            proxyHeaders: {
-                'Access-Control-Allow-Origin': "https://proxy.hackeryou.com"
-            }
-        },
+//             proxyHeaders: {
+//                 'Access-Control-Allow-Origin': "https://proxy.hackeryou.com"
+//             }
+//         },
 
-        xmlToJSON: false
+//         xmlToJSON: false
 
-}).then ((res) => {
-// Convert XML to JS Object
-    const userSearchResults = [];
+// }).then ((res) => {
+// // Convert XML to JS Object
+//     const userSearchResults = [];
 
-    const result3 = convert.xml2js(res.data, {compact: false, spaces: 2});
+//     const result3 = convert.xml2js(res.data, {compact: false, spaces: 2});
     
-    const condensedRes2 = result3.elements[0].elements[1].elements[6].elements;
-
-    
-    
-    const userSearchRes = [...condensedRes2];
-    console.log(userSearchRes);
-
-// Push search results into empty array
-    userSearchRes.map(book => {
-
-        return (
-            userSearchResults.push({
-                title: book.elements[8].elements[1].elements[0].text,
-
-                author: book.elements[8].elements[2].elements[1].elements[0].text === undefined 
-                ? "" 
-                : book.elements[8].elements[2].elements[1].elements[0].text,
-
-                key: book.elements[8].elements[0].elements[0].text,
-
-                rating: book.elements[7].elements[0].text,
-
-                imageUrl: book.elements[8].elements[3].elements[0].text === undefined ? "/src/assets/noCover.jpg"
-                : book.elements[8].elements[3].elements[0].text,
-            })
-        );
-    })
-
-// Store searched books
-    this.setState({
-        books: userSearchResults,
-    })
-    console.log(userSearchResults);
-
-
+//     const condensedRes2 = result3.elements[0].elements[1].elements[6].elements;
 
     
-})
+    
+//     const userSearchRes = [...condensedRes2];
+//     console.log(userSearchRes);
 
-// Clear search input
-    this.setState({
-        searchInput: " ",
+// // Push search results into empty array
+//     userSearchRes.map(book => {
 
-    })
+//         return (
+//             userSearchResults.push({
+//                 title: book.elements[8].elements[1].elements[0].text,
+
+//                 author: book.elements[8].elements[2].elements[1].elements[0].text === undefined 
+//                 ? "" 
+//                 : book.elements[8].elements[2].elements[1].elements[0].text,
+
+//                 key: book.elements[8].elements[0].elements[0].text,
+
+//                 rating: book.elements[7].elements[0].text,
+
+//                 imageUrl: book.elements[8].elements[3].elements[0].text === undefined ? "/src/assets/noCover.jpg"
+//                 : book.elements[8].elements[3].elements[0].text,
+//             })
+//         );
+//     })
+
+// // Store searched books
+//     this.setState({
+//         searchedBooks: userSearchResults,
+//     })
+//     console.log(userSearchResults);
+   
+// })
+
+// // Clear search input
+//     this.setState({
+//         searchInput: " ",
+
+//     })
 
 
-}
+// }
   
   
 
@@ -206,12 +206,13 @@ class App extends Component {
         {/* Firebase Setup */}
         <Header />
 
-        <Main/>
-        <Footer/>
-
-        <SearchResults />
+        <Main />
+        
+        <Footer />
+        
         
 
+    
       </div>
     );
   }
