+++ src/Components/Header.js
@@ -1,23 +1,18 @@
-import React, { Component } from 'react';
+import React from 'react';
 
-class Header extends Component {
+const Header = () => {
 
-    constructor (){
-        super();
-    }
+    return(
 
-    render (){
-        return (
-            <header>
+        <header>
 
-                <h1>Lit Tracker</h1>
+            <h1>Lit Tracker</h1>
 
-            </header>
-            // stuff 
-        )
-    }
+        </header>
 
-  } 
+    )   
+   
+}
